Create io lib
