+++ tcp.go
@@ -1,10 +1,9 @@
 package main
 
 import (
-	"io"
 	"net"
 
-	"github.com/shadowsocks/go-shadowsocks2/shadowaead"
+	"github.com/shadowsocks/go-shadowsocks2/io"
 	"github.com/shadowsocks/go-shadowsocks2/socks"
 )
 
@@ -80,7 +79,7 @@ func tcpLocal(addr, server string, shadow func(net.Conn) net.Conn, getAddr func(
 			}
 
 			logf("proxy %s <-> %s <-> %s", c.RemoteAddr(), server, tgt)
-			_, _, err = relay(c, rc)
+			_, _, err = io.Relay(c, rc)
 			if err != nil {
 				logf("relay error: %v", err)
 			}
@@ -125,43 +124,10 @@ func tcpRemote(addr string, shadow func(net.Conn) net.Conn) {
 			rc.(*net.TCPConn).SetKeepAlive(true)
 
 			logf("proxy %s <-> %s", c.RemoteAddr(), tgt)
-			_, _, err = relay(c, rc)
+			_, _, err = io.Relay(c, rc)
 			if err != nil {
 				logf("relay error: %v", err)
 			}
 		}()
 	}
-}
-
-// relay copies between left and right bidirectionally. Returns number of
-// bytes copied from right to left, from left to right, and any error occurred.
-func relay(left, right io.ReadWriteCloser) (int64, int64, error) {
-	type res struct {
-		N   int64
-		Err error
-	}
-	ch := make(chan res)
-
-	go func() {
-		n, err := io.Copy(right, left)
-		shadowaead.CloseWrite(right)
-		if n == 0 || err != nil {
-			shadowaead.CloseRead(right)
-		}
-		logf("copyClose L->R done: %v %v", n, err)
-		ch <- res{n, err}
-	}()
-
-	n, err := io.Copy(left, right)
-	shadowaead.CloseWrite(left)
-	if n == 0 || err != nil {
-		shadowaead.CloseRead(left)
-	}
-	logf("copyClose L<-R done: %v %v", n, err)
-	rs := <-ch
-
-	if err == nil {
-		err = rs.Err
-	}
-	return n, rs.N, err
 }
